pyproject.toml
[project]
name = "batch-agent"
version = "0.1.0"
description = "Batch Agent SDK foundation for high-throughput agent orchestration."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2",
]

[project.optional-dependencies]
cli = ["pyyaml>=6"]
test = ["pytest>=8", "pytest-asyncio>=0.23"]

[project.scripts]
batch-agent = "batch_agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
